Alterar Produto Teste de Browser

+++ app/Http/Controllers/DeletarProdutoController.php
@@ -7,6 +7,7 @@ use App\Validator\ProdutoValidator;
 use App\Validator\ValidationException;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Validator;
 
 class deletarProdutoController extends Controller
@@ -15,7 +16,9 @@ class deletarProdutoController extends Controller
     public function deletar(Request $request)
     {
             $produto = Produto::find($request->id);
+        if(Gate::allows('delete-produto', $produto)) {
             $produto->delete();
+        }
             return redirect('/listarProdutos');
 
     }
